% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_form.R
\name{p_form}
\alias{p_form}
\title{Format p-values Function}
\usage{
p_form(x, lead = F, stars = T)
}
\arguments{
\item{lead}{Should leading zeroes be shown? FALSE}

\item{stars}{Should stars/asterisks be added? TRUE}
}
\description{
This function allows you to format p-values w/o stars and leading/trailing zeroes.
}
\examples{
p_form()
}
\keyword{pvalues}
