% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_form_proxy.R
\name{p_form_proxy}
\alias{p_form_proxy}
\title{Format arbitrary values Function}
\usage{
p_form_proxy(pval, target, lead = T, digits = 2)
}
\arguments{
\item{lead}{Should leading zeroes be shown? FALSE}

\item{digits}{How many digits should the target value be rounded to? 2}
}
\description{
This function allows you to add asterisks to arbitrary values; conditional on p-values.
}
\examples{
p_form_proxy()
}
\keyword{asterisks}
