% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabs.R
\name{reliabs}
\alias{reliabs}
\title{Reliability Function}
\usage{
reliabs(dat, fm = "minres", nfactors = 1)
}
\arguments{
\item{dat}{Dataframe containing the relevant variables}

\item{fm}{Method for factor structure estimation (default: minres)}

\item{nfactors}{Number of factors to estimate omega for (default: 1)}
}
\description{
This function calculates different estimates for reliability/internal consistency coefficients.
}
\examples{
reliabs()
}
\keyword{reliability}
