% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundy.R
\name{roundy}
\alias{roundy}
\title{Round Function}
\usage{
roundy(x, digits = 0, lead = T)
}
\arguments{
\item{digits}{Number of digits to show 0}

\item{lead}{Should leading zeroes be shown? TRUE}
}
\description{
This function allows you to round numbers more easily w/o leading and trailing zeroes.
}
\examples{
roundy()
}
\keyword{round}
